Add fuzz_one_pool_test
